+++ BuildCraftAPI/api/buildcraft/api/tiles/TilesAPI.java
@@ -2,22 +2,24 @@ package buildcraft.api.tiles;
 
 import javax.annotation.Nonnull;
 
-import net.neoforged.neoforge.capabilities.Capability;
+import net.minecraft.core.Direction;
+
+import net.neoforged.neoforge.capabilities.BlockCapability;
 
 import buildcraft.api.core.CapabilitiesHelper;
 
 public class TilesAPI {
     @Nonnull
-    public static final Capability<IControllable> CAP_CONTROLLABLE;
+    public static final BlockCapability<IControllable, Direction> CAP_CONTROLLABLE;
 
     @Nonnull
-    public static final Capability<IHasWork> CAP_HAS_WORK;
+    public static final BlockCapability<IHasWork, Direction> CAP_HAS_WORK;
 
     @Nonnull
-    public static final Capability<IHeatable> CAP_HEATABLE;
+    public static final BlockCapability<IHeatable, Direction> CAP_HEATABLE;
 
     @Nonnull
-    public static final Capability<ITileAreaProvider> CAP_TILE_AREA_PROVIDER;
+    public static final BlockCapability<ITileAreaProvider, Direction> CAP_TILE_AREA_PROVIDER;
 
     static {
         CAP_CONTROLLABLE = CapabilitiesHelper.registerCapability(IControllable.class);